BIP Claims Evidence Final Integration (#992)
bip ce final integration
