+++ service/provider/src/main/java/gov/va/vro/service/provider/bip/service/BipClaimService.java
@@ -18,12 +18,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ObjectUtils;
 import org.springframework.stereotype.Service;
 
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.StandardCopyOption;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.Collection;
@@ -40,6 +35,8 @@ public class BipClaimService {
   public static final String SPECIAL_ISSUE_2 = "rrd";
 
   private final IBipApiService bipApiService;
+
+  private final IBipCeApiService bipCeApiService;
 
   /**
    * Check if all the anchors for fast-tracking are satisfied.
@@ -182,51 +179,29 @@ public class BipClaimService {
     String filename =
         GeneratePdfPayload.createPdfFilename(
             DiagnosisLookup.getDiagnosis(payload.getDiagnosticCode()));
-    File file = null;
-    try {
-      file = File.createTempFile(filename, "tmp", null);
-      byte[] decoder = Base64.getDecoder().decode(pdfResponse.getPdfData());
-      InputStream is = new ByteArrayInputStream(decoder);
-      Files.copy(is, file.toPath(), StandardCopyOption.REPLACE_EXISTING);
-      List<String> contentionList = new ArrayList<>(); // find contention related
-      BipFileProviderData providerData =
-          BipFileProviderData.builder()
-              .contentSource("VRO")
-              .claimantFirstName(payload.getFirstName())
-              .claimantLastName(payload.getLastName())
-              .claimantSsn(payload.getVeteranIdentifiers().getSsn())
-              .benefitTypeId(10)
-              .documentTypeId(131)
-              .claimantDateOfBirth("1900-01-01")
-              .dateVaReceivedDocument("1900-01-01") // don't know what data is
-              .subject(pdfResponse.getDiagnosis()) // get a subject
-              .contentions(contentionList)
-              .alternativeDocumentTypeIds(List.of(1))
-              .actionable(false)
-              .associatedClaimIds(List.of("1"))
-              .notes(pdfResponse.getReason() == null ? null : pdfResponse.getReason())
-              .payeeCode("00")
-              .endProductCode("130DPNDCY")
-              .regionalProcessingOffice("Buffalo") // get an office.
-              .facilityCode("Facility")
-              .claimantParticipantId(payload.getVeteranIdentifiers().getParticipantId())
-              .sourceComment("upload from VRO")
-              .claimantDateOfBirth(payload.getDateOfBirth())
-              .build();
 
-      bipApiService.uploadEvidence(
-          FileIdType.FILENUMBER,
-          pdfResponse.getClaimSubmissionId(),
-          BipFileUploadPayload.builder().contentName(filename).providerData(providerData).build(),
-          file);
-      return pdfResponse;
-    } catch (IOException ioe) {
-      throw new BipException("Failed to upload evidence file.", ioe);
-    } finally {
-      if (file != null) {
-        file.delete();
-      }
-    }
+    byte[] decoder = Base64.getDecoder().decode(pdfResponse.getPdfData());
+    BipFileProviderData providerData =
+        BipFileProviderData.builder()
+            .contentSource("VRO")
+            .claimantFirstName(payload.getFirstName())
+            .claimantLastName(payload.getLastName())
+            .claimantSsn(payload.getVeteranIdentifiers().getSsn())
+            .documentTypeId(1489)
+            .dateVaReceivedDocument(LocalDate.now().toString())
+            .subject(pdfResponse.getDiagnosis()) // get a subject
+            .notes(pdfResponse.getReason() == null ? null : pdfResponse.getReason())
+            .claimantParticipantId(payload.getVeteranIdentifiers().getParticipantId())
+            .sourceComment("upload from VRO")
+            .claimantDateOfBirth(payload.getDateOfBirth())
+            .build();
+
+    bipCeApiService.uploadEvidenceFile(
+        FileIdType.FILENUMBER,
+        payload.getVeteranIdentifiers().getVeteranFileId(),
+        BipFileUploadPayload.builder().contentName(filename).providerData(providerData).build(),
+        decoder);
+    return pdfResponse;
   }
 
   private static boolean hasSpecialIssues(ClaimContention claimContention) {
